+++ src/components/App.js
@@ -1,9 +1,16 @@
 import Posts from "./Posts";
+import PostForm from "./PostForm";
+import { Route, BrowserRouter, Switch } from "react-router-dom";
 
 function App() {
   return (
     <main>
-      <Posts />
+      <BrowserRouter>
+        <Switch>
+          <Route exact path="/" component={Posts} />
+          <Route path="/new" component={PostForm} />
+        </Switch>
+      </BrowserRouter>
     </main>
   );
 }
