+++ src/components/Posts.js
@@ -1,25 +1,11 @@
 import Post from "./Post.js";
-import PostForm from "./PostForm.js";
-import { useState } from "react";
-
-function getPosts() {
-  const PostsString = localStorage.posts || "[]";
-  const Posts = JSON.parse(PostsString);
-  return Posts;
-}
+import { getPosts } from "../utils/postsStorage.js";
 
 function Posts() {
-  const [posts, setPosts] = useState(getPosts);
-
-  function handleAddPost(newPost) {
-    const newPosts = [...posts, newPost];
-    localStorage.setItem("posts", JSON.stringify(newPosts));
-    setPosts(newPosts);
-  }
+  const posts = getPosts();
 
   return (
     <>
-      <PostForm onAddPost={handleAddPost} />
       {posts.map((post) => (
         <Post {...post} key={post.id} />
       ))}
